Refactored my tasks for the weeks
