+++ src/sZahir/Week01/FINRA.java
@@ -1,4 +1,4 @@
-package sZahir;
+package sZahir.Week01;
 
 public class FINRA {
