+++ src/sZahir/Week02/Codility.java
@@ -1,4 +1,4 @@
-package sZahir;
+package sZahir.Week02;
 
 /*
 Write a function that given a positive integer N prints the consecutive numbers
